Обновление классов и тестов для работы с SKU
Добавлены новые классы и методы для работы с наборами мерча (SkuPreset) и их типами (PresetType). Изменён тип идентификатора в Entity с int на long для поддержки больших значений. Обновлены тесты для проверки функциональности SkuPreset. Изменена структура проекта и добавлены необходимые зависимости для тестирования с использованием MSTest и xUnit.

+++ OzonEdu.MerchandiseService.Domain/Root/Entity.cs
@@ -19,7 +19,7 @@ namespace OzonEdu.MerchandiseService.Domain.Root
 		/// <summary>
 		/// Идентификатор сущности
 		/// </summary>
-		public virtual int Id { get; protected set; }
+		public virtual long Id { get; protected set; }
 
 		/// <summary>
 		/// Список доменных событий
